RequirementMachine: Split off RewriteContext.{h,cpp} from RewriteSystem.{h,cpp}
